add accepting transaction
