+++ PayServer/src/main/resources/ClientsInfo
@@ -1,4 +1,4 @@
-steps stst 1 100
-msh mmm 2 10 
-kek kk 3 350
-igor 1111 945763 0
+s 1 40 
+m 2 110 
+qwe 363981 120 
+asd 695726 900
